backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
openai==1.3.6
stripe==7.0.0
redis==4.6.0
psycopg2-binary
prisma==0.10.0